backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
chromadb==0.5.0
pydantic==2.9.0
python-dotenv==1.0.0
httpx==0.27.0
pytest==8.3.0
pytest-cov==5.0.0
